libui: Change hasSignaled to return optional<bool>
Since the android::Fence::wait() call can possibly return an error,
we need some way of indicating to the caller of hasSignaled that a
fence that has not yet signaled (because it is in an error state) will
never signal. To do this, we return an optional<bool>, where true and
false indicate both that the fence is valid and that it has or hasn't
signaled. If an error is returned from wait(), we return a default
optional value (that is neither true nor false).

Test: m

Change-Id: Ibce48cd2e71ddb8ccf6cabe3284afe0efca8c132