disable explicit gc to avoid app slowdown
Proprietary :( apps like instagram constantly hangs because their devs
are calling the gc unnecessarily from the UI thread. Making the explicit
gc a noop avoids the lag and the app remains perfectly fluid and
interactive at all times.

Without this change instagram is basically unusable on Android.

I've been running with this change for a couple of days with no
regression to any app and no memleak either.

This is fundamentally a tradeoff between overall higher memory
utilization and speed. I prefer speed than keeping around more
background tasks.

Clearly it's a bug in the app, but in general it's very rarely good idea
to call gc explicitly and the downside of not calling it is much less
concerning than the downside if the app is calling it too much.

Change-Id: I289c86ba621c02c7e6084a4c4909a7544dc5be1e