Update patch set 1
Patch Set 1:

Hello,

in the commit message I didn't specify I consider this more a RFC for now. If apps were optimal it wouldn't be necessary. However apps may be using libs that may be calling the gc too for whatever reason, without the app developer even realizing it.

To make a test I tried to disable it and things eventually got very slow again. The very noticeable scrolling lag resurfaced. So I had to enable it again to retain perfect interactivity at all times while using the app. If this gets applied it should be tested for a while and see how it goes before it can go in a stable release.

I know I need to keep it applied here.

The best would be to be able to tweak this at runtime on a per-app basis (with global option of explicit gc disabled/enabled and then having the apps opt-in/opt-out). Or alternatively (less finegrined) with a single global option in the debug menu.

About moving this tweak to a specific device I don't see why, I doubt the app checks for a specific device to decide if to call the gc. If something this should be extended to the larger mem configuration file.

This is an highlevel API issue that shouldn't be decided on a single app (which is why a RFC and more testing is more appropriate for now) and so I doubt it should be done on a single device either.

This is a valid solution for whoever else may run into scrolling lag or bad performance caused by this. Frankly I've been surprised that this solved my problem completely and that the app was really doing so many unnecessary gc calls. It's an app with 1 billion downloads on google play. My rationale for applying this globally and not just for such app, is that if they can't get it right, probably they're not the only ones doing such optimization mistakes and relaying on JVM intelligence to do the gc is probably better overall. I guess this is also why the option of ignoring explicit gc exists in the art command line in the first place.

Patch-set: 1
Reviewer: Andrea Arcangeli <3864@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0