Update patch set 1
Patch Set 1:

I would guess this isn't occurring on all devices, which is why this is better off in the device repository. Also, this is only one of several VM configs, not all devices use this. If the goal is to apply it globally, it might be better to move to vendor/cm instead.

I don't dispute that this improves your app experience though.

Patch-set: 1