Update patch set 1
Patch Set 1: Code-Review+1

> the flag is already non-existent. Besides, it isn't the right way
 > to fix missing symbols - shims are meant for this.

If you take the commit by steve which adds the m-compatible sensors abi and then pick these commits I sent in for you, it'll work. I tested it on my own device and it is working perfectly fine. No more shimming or hexpatching of proprietary vendor blobs!! And this can be used on other devices as well.

Patch-set: 1