Update patch set 1
Patch Set 1: Code-Review-2

The issue is that you're actually changing the ABI in other potentially incompatible ways that we can't really predict. Shims *are* the right way to do this because they allow you to limit the scope of the change to a single component. I wrote the original version of this patch and quickly realized that it was a bad idea.

Please *don't* include this in your builds. Use the shim I wrote for the OP3 instead.

Patch-set: 1
Label: Code-Review=-2