sf: hwc1: Set SF EventThread to SCHED_FIFO in all cases
Previously, we were only setting the SF EventThread to SCHED_FIFO
when the phase offsets aren't the same. This would probably result
in more jitter for devices whose phase offsets are the same.

This was a result of the Androdi 7.1 mismerge.

This essentially reverts commit 12f58910e54084339ae794af7652d4b0dc3818b0
and fixes the Android 7.1 merge.

Change-Id: I73594001e381c176410e06865352dff74019ff27