vulkan: Correct access after free issue
While testing argument validation in the loader, I found that
I got SEGFAULTs trying to access the logger class created in
LayerChain::CreateDevice. I believe that the logger object
was being destroyed at the end of the LayerChain construction
statement so that a reference to the logger caused a SEGFAULT.

Fix CreateInstance path as well, though the loader doesn't
currently call the logger in that path, so no way to verify.

Test: run vkCreateDevice_test on Vulkan device.
See tests/README.md for details
Bug: 33783620

Change-Id: Ib46023e8a8dfa008510c655c5b216aabaaf2ea83