Enumerate initial version of VK_GOOGLE_display_timing
The implementation seems correct with the exception that it is not obtaining
the refresh durations from SurfaceFlinger/HAL.  A maximum of 10 TimingInfo's
(containing the timing for previous presents) are stored, and timestamps are
only queried from SurfaceFlinger for presents (a.k.a. frames) that occured 5
presents/frames ago.

Test: Manually tested with a modified cube demo, that changes its timing as a
result of the data returned from this extension.

Change-Id: I199614ed58877b64eab2568a76ba517acca3bf87