Fix -Wdouble-promotion warnings
This is a manual merge of change
https://android-review.googlesource.com/#/c/328520/ , since automerging
failed.

With the new clang FORTIFY, this code emits the following error:

error: implicit conversion increases floating-point precision: 'float'
to 'double' [-Werror,-Wdouble-promotion]

This is because strncpy is now a function, not a macro defined in a
system header. So, the warning is considered to appear in user code
(instead of system code), and is therefore no longer suppressed.

Bug: 32073964
Test: Now builds with clang FORTIFY
Change-Id: I5e8292d7800434268793bc4db4a96f20b7022e17