Preserve app data if neither ce nor de have xattr user.default
For a given app, if xattr user.default is missing from both ce and de data
dirs, installd currently defaults to using ce.  However, if a system
app sets defaultToDeviceProtectedStorage="true" in it's manifest,
data is actually stored in de and so the effect is wipe for such
apps in the case where xattr user.default attributes are missing
(such as a tar backup/restore).

In the case where user.default is missing, this patch checks both
ce and de dirs for the presence of data and will use de in the case
where ce is empty and de is not.  All other cases fall through
to the original behaviour of using ce by default.

Change-Id: I8d5583b7d156809b2aecc676db35e6bf7cad3e0c