Ignore VK_KHR_incremental_present hint for MAILBOX.
Test: Using instrumented cube demo, and tmp ALOGI comments.

This code may eventually be replaced with code that constructs, stores,
and uses regions that spans multiple vkQueuePresentKHR() calls.  With
VK_PRESENT_MODE_MAILBOX_KHR, any given present may not be sent to the
display.  In such a case, the correct behavior is for the swapchain code
(a.k.a. acting for the "presentation engine") should build up a list of
regions across multiple vkQueuePresentKHR() calls, so that when an image
is presented, all of the updated regions will be sent to the display.
In the mean time, this code is correct (i.e. ignore the hint for
MAILBOX).

Change-Id: I413091871a9b9b7464ec941a857eea56e68f4189