Update patch set 1
Patch Set 1:

> i'm fairly certain the current plan is cmte

On the theme commit, it isn't explicity said that we are still using CMTE. As said by Clark, we can submit the patches and see where it goes from there. If the final choice is to go to CMTE, then I can either stop uploading the last patches for framework/base or continue and abandon everything.

Patch-set: 1