vulkan: Split known/intercepted extension lists
We need to do some work in the loader based on whether gpdp2 is present
and enabled, and we'd like to /not/ filter gpdp2 out of extension lists.

However, we don't need or want to generate full forwarding stubs.

All extensions in driver.KnownExtensions will have enums and matching
infrastructure generated, but only extensions in
driver.InterceptedExtensions will have g_hook_procs populated for their
entrypoints.

V3: Define driver.KnownExtensions in terms of
driver.InterceptedExtensions as always a superset.

Change-Id: If0fdabad99fa4637d7c6fc1e9a7e5e3908b53aca
Test: build