VR: Replace sw sync fences with egl sync fences
HWComposer::presentDisplay() synchronously calls into the VR WM
presentation code to queue the new frame. If VR WM is showing (drawing)
it will return an EGL fence that will be used to signal when the last
showing frame is no longer used in VR WM. Otherwise an invalid FD (-1)
is returned symbolizing that the last frame isn't in use anymore and the
caller can re-use those buffers immediately.

Bug: b/35096352
Test: Verified SurfaceFlinger output shows correctly in vr_wm
Change-Id: I2b7c2c767d4aa85370dd6519113cb9a2269c7477