Squashed merge of master-sim into master.
Includes the following commits:

==

Add data file for new eUICC hardware feature.

A device may declare hardware support for eUICC APIs by copying this
file to vendor/etc/permissions/. Doing so is a declaration that all
eUICC APIs are expected to function (and pass CTS), and that apps
which require eUICC support to function are eligible to be installed.

==

Test: TreeHugger
Change-Id: Ia7f5f9925c51857fd8034b8c882988cb1b731ff2