Update patch set 1
Patch Set 1:

Providing you leave the Change-Id in your commit messages and don't mess with it, it will upload a new patchset here, rebased on the framework_base changes.

The changes merged were for android_7.1.1-r21 (the February security patches).

I've repopicked this topic and merge conflicted as best I can. The conflicts are not as bad as they first seem :)

Patch-set: 1