[surfaceflinger] Set blend mode correctly when using HWC2.
HWC1 code will reset blend mode to None using setDefaultState() but
HWC2 just leaves the layer in the previous state. The result is that
once blend mode is set to something other than HWC2::BlendMode::None
we can never go back to HWC2::BlendMode::None and HWC2 implementations
are forced to use blending when not necessary.

Bug: 35915184
Test: manually verified using overdraw feedback on arc enabled chromebook
Change-Id: I86957439bad0a1be9faa56df491d8415c4cdd2c2