Add a simple SurfaceFlinger stress test.
Obviously such tests are always a little arbitrary but this one
seems pretty useful. It runs in ~2 seconds, I got 10 passes in a row with
the fix for 36117490 applied, and 3 system restarts in a row (before giving up)
without the fix applied. Clean up the other tests in preparation for presubmit enabling.

Test: Quis custodiet ipsos custodes?
Bug: 36117490
Change-Id: Ifef2ff4e51aa9af193ddf003fe4aba053b28738f