vulkan: Always request at least 2 images for swapchain
For shared mode, we really only care about 1 image, but there are other
changes further down the stack that need to be made for that to really
work. For now, just tell the bufferqueue that we want 2. After that, we
still only dequeue one buffer (as in fact required by BQ in shared mode)
and construct one VkImage over it for giving back to the vulkan app.

[in V1 we were just smashing the buffer count to 2, which is messier and
observable by the app]

Change-Id: Idc0d3c232d56a5c0127f9472611d151f3605939c
Test: build