media: mechanism to set/get vendor-extension configs for OMX component
Add OMX_CONFIG_ANDROID_VENDOR_EXTENSIONTYPE and
OMX_CONFIG_ANDROID_VENDOR_PARAMTYPE.

Android vendor extensions for OMX components have a name and one or
more parameter values - each with a string key - that are set together.
Support for not-set values is provided for vendor extensions in order
to not suddenly confuse applications with format keys that they were
not expecting.

Bug: 32746650
Change-Id: I21aad4a82d509fe2466764f65ee9e3b2d32f0842