Disable BufferQueueInAnotherProcess test
The BufferQueueInAnotherProcess relies on executing before any test that
tries to open a BufferQueue in the local process. A recent test addition
broke the condition. Upcoming SF & BQ changes are likely to obsolete the
test, so won't spend time on fixing it. Disabling for now and the test
will be removed when the infrastructure changes land.

Test: libgui_test
Bug: 36592665
Change-Id: I648e197cab6bab3aed445390069265589a717bf1